+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+- [csvtk v0.31.2](https://github.com/shenwei356/csvtk/releases/tag/v0.31.2)
+[![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.31.2/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.31.2)
+    - `csvtk filter2/mutate2/mutate3`:
+        - fix a bug of mismatch between column names and values. [#295](https://github.com/shenwei356/csvtk/issues/295)
 - [csvtk v0.31.1](https://github.com/shenwei356/csvtk/releases/tag/v0.31.1)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.31.1/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.31.1)
     - `csvtk filter2/mutate2/mutate3`: